dropship_automator_scaffold/apps/worker/requirements.txt
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
requests==2.32.3
